# docs/swagger.yaml
basePath: /api/
definitions:
...
  db.Materia:
    properties:
      codigo:
        type: string
      creditos:
        type: integer
      descripcion:
        $ref: '#/definitions/pgtype.Text'
      facultad:
        type: string
      materiaID:
        type: integer
      nombre:
        type: string
    type: object
...
  handler.CreateMateriaRequest:
    properties:
      codigo:
        example: MATH101
        type: string
      creditos:
        example: 4
        type: integer
      descripcion:
        example: Introducción al cálculo diferencial e integral
        type: string
      facultad:
        example: Ingeniería
        type: string
      nombre:
        example: Cálculo I
        type: string
    type: object
  handler.CreateMateriaResponse:
    properties:
      materia_id:
        type: integer
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: Descriptive error message
        type: string
    type: object
...
  handler.UpdateMateriaRequest:
    properties:
      codigo:
        example: MATH101
        type: string
      creditos:
        example: 4
        type: integer
      descripcion:
        example: Introducción al cálculo diferencial e integral
        type: string
      facultad:
        example: Ingeniería
        type: string
      nombre:
        example: Cálculo I
        type: string
    type: object
...
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
...
host: matwa.tail013c29.ts.net
info:
  contact: {}
  description: This is the API for my project.
  termsOfService: http://swagger.io/terms/
  title: Proyecto de Ingenieria de  Datos API
  version: "1.0"
paths:
...
  /v1/materias:
    get:
      description: Retrieves materias filtered by faculty.
      parameters:
      - description: Faculty name
        in: query
        name: facultad
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved materias
          schema:
            items:
              $ref: '#/definitions/db.Materia'
            type: array
        "500":
          description: Failed to retrieve materias
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List Materias by Faculty
      tags:
      - Materias
    post:
      consumes:
      - application/json
      description: Creates a new materia (subject).
      parameters:
      - description: Materia Data
        in: body
        name: materia
        required: true
        schema:
          $ref: '#/definitions/handler.CreateMateriaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created materia
          schema:
            $ref: '#/definitions/handler.CreateMateriaResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to create materia
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create Materia
      tags:
      - Materias
  /v1/materias/{id}:
    delete:
      description: Deletes a materia by its ID.
      parameters:
      - description: Materia ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted materia
        "400":
          description: Invalid materia ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to delete materia
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Materia
      tags:
      - Materias
    get:
      description: Retrieves a specific materia by its ID.
      parameters:
      - description: Materia ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved materia
          schema:
            $ref: '#/definitions/db.Materia'
        "400":
          description: Invalid materia ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Materia not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to retrieve materia
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get Materia by ID
      tags:
      - Materias
    put:
      consumes:
      - application/json
      description: Updates an existing materia.
      parameters:
      - description: Materia ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Materia Data
        in: body
        name: materia
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateMateriaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated materia
          schema:
            $ref: '#/definitions/db.Materia'
        "400":
          description: Invalid request body or materia ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Materia not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to update materia
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update Materia
      tags:
      - Materias
  /v1/materias/codigo/{codigo}:
    get:
      description: Retrieves a specific materia by its code.
      parameters:
      - description: Materia Code
        in: path
        name: codigo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved materia
          schema:
            $ref: '#/definitions/db.Materia'
        "404":
          description: Materia not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to retrieve materia
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get Materia by Code
      tags:
      - Materias
...
schemes:
- https
swagger: "2.0"